changing the side by name
